Move opcodes to their corresponding folder (#863)
